Refactoring: loginEmail 페이지 에러 수정 및 디자인패턴 적용
